split demo code into multiple files
