--- sessions/js-array-methods-2/demo-end/js/index.test.js
@@ -1,6 +1,0 @@
-import { myFunction } from "./index.js";
-
-test('myFunction() returns "It works!"', () => {
-  const result = myFunction();
-  expect(result).toBe("It works!");
-});
